Add condition on removing column in BookingTables migration

+++ db/migrate/20180831165554_change_status_id_column_on_booking_tables.rb
@@ -1,9 +1,13 @@
 class ChangeStatusIdColumnOnBookingTables < ActiveRecord::Migration[5.2]
   def change
-    remove_column :active_books, :status_id
+    if column_exists? :active_books, :status_id
+      remove_column :active_books, :status_id
+    end
     add_column :active_books, :order_status_id, :integer
 
-    remove_column :book_histories, :status_id
+    if column_exists? :book_histories, :status_id
+      remove_column :book_histories, :status_id
+    end
     add_column :book_histories, :order_status_id, :integer
   end
 end
